introduce target point controller
